Add files via upload
The first 4 commits of our project. This composes of the the data cleaning and preparation as well as exploratory data analysis. This was all work from semester one.  We were able to continue the project from here individually, where each of us could focus on our own machine learning algorithm going forward.